refactor(Knowledge Dashboard): 合并 Dashboard 与 Pipelines 页面为统一的融合视图;
将 Knowledge 模块下的 Dashboard 页与 Pipelines 页合并为一个增强型 Dashboard 页，
消除 Pipeline Runs 与 Runs 列表的功能重叠，并将 Timeline 信息内嵌到融合卡片中。

主要变更：
- PipelineRunCard 新增 selectable 模式，支持选中高亮与 Timeline 时间戳内联展示
- 提取 PipelineRunDetailPanel 为独立组件（Info/Stages/Input/Output/Errors）
- Dashboard 页并行加载 fetchDashboard + fetchPipelines，迁移三层轮询逻辑
- 移除导航栏 Pipelines 入口，原路由改为 redirect 兜底
- 测试迁移至 DashboardPage.test.tsx，清理 Pipelines 截图基线
