+++ apps/negentropy-ui/components/ui/KnowledgeNav.tsx
@@ -10,7 +10,6 @@ const NAV_ITEMS = [
   { href: "/knowledge/base", label: "Knowledge Base" },
   { href: "/knowledge/documents", label: "Documents" },
   { href: "/knowledge/graph", label: "Knowledge Graph" },
-  { href: "/knowledge/pipelines", label: "Pipelines" },
   { href: "/knowledge/apis", label: "APIs" },
 ];
